feature: use bloc pattern for artist agenda component

+++ lib/ui/on_boarding/widgets/custom_dots_indicator.dart
@@ -20,7 +20,7 @@ class CustomDotsIndicator extends StatelessWidget {
               padding: const EdgeInsets.only(left: 39),
               child: DotsIndicator(
                 dotsCount: state.contents.length,
-                position: state.page.toDouble(),
+                position: state.page,
                 mainAxisAlignment: MainAxisAlignment.start,
                 onTap: (index) {
                   context
